backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
boto3==1.29.3
sqlalchemy==1.4.51
psycopg2-binary==2.9.10
alembic==1.12.1 